+++ app/src/main/java/com/mtkreader/contracts/ReadingContract.kt
@@ -6,6 +6,9 @@ import android.content.Context
 import androidx.fragment.app.Fragment
 import com.mtkreader.commons.base.AutoDisposePresenter
 import com.mtkreader.commons.base.ErrorHandlingFragment
+import com.mtkreader.data.DeviceDate
+import com.mtkreader.data.DeviceTime
+import io.reactivex.subjects.PublishSubject
 
 interface ReadingContract {
 
@@ -15,7 +18,8 @@ interface ReadingContract {
         fun onSocketConnected(socket: BluetoothSocket)
         fun onReceiveBytes(byte: Byte)
 
-        fun displayTimeData(time:String)
+        fun displayTimeData(time: String)
+        fun onTimeWriteResult(isSuccessful: Boolean)
 
         fun onError(throwable: Throwable)
     }
@@ -25,5 +29,7 @@ interface ReadingContract {
         fun readStream(socket: BluetoothSocket)
         fun closeConnection()
         fun extractTimeData(context: Context, data: List<Char>, hardwareVersion: Int)
+        fun setTimeDate(time: DeviceTime, deviceDate: DeviceDate)
+        fun setData(data: List<Char>)
     }
 }
