+++ app/src/main/java/com/mtkreader/views/fragments/ReadingView.kt
@@ -15,20 +15,11 @@ import com.ikovac.timepickerwithseconds.MyTimePickerDialog
 import com.ikovac.timepickerwithseconds.TimePicker
 import com.mtkreader.R
 import com.mtkreader.commons.Const
-import com.mtkreader.commons.Const.Data.ACK
-import com.mtkreader.commons.Const.Data.COMPLETE
-import com.mtkreader.commons.Const.Data.EOT
-import com.mtkreader.commons.Const.Data.ETX
-import com.mtkreader.commons.Const.Data.NAK
-import com.mtkreader.commons.Const.Data.SOH
-import com.mtkreader.commons.Const.Data.STX
 import com.mtkreader.commons.base.BaseMVPFragment
 import com.mtkreader.contracts.ReadingContract
 import com.mtkreader.data.DeviceDate
 import com.mtkreader.data.DeviceTime
-import com.mtkreader.data.reading.TimeDate
 import com.mtkreader.data.writing.DataRXMessage
-import com.mtkreader.data.writing.DataTXTMessage
 import com.mtkreader.presenters.ReadingPresenter
 import com.mtkreader.utils.CommunicationUtil
 import com.mtkreader.utils.DataUtils.getHardwareVersion
@@ -36,10 +27,8 @@ import com.mtkreader.utils.SharedPrefsUtils
 import com.mtkreader.utils.TimeUtils
 import com.mtkreader.views.dialogs.ConnectingDialog
 import kotlinx.android.synthetic.main.fragment_reading.*
-import java.util.*
-import kotlin.concurrent.thread
-import kotlin.experimental.or
-import kotlin.experimental.xor
+import net.alexandroid.utils.mylogkt.logD
+import net.alexandroid.utils.mylogkt.logI
 
 class ReadingView : BaseMVPFragment<ReadingContract.Presenter>(), ReadingContract.View,
     MyTimePickerDialog.OnTimeSetListener {
@@ -81,7 +70,7 @@ class ReadingView : BaseMVPFragment<ReadingContract.Presenter>(), ReadingContrac
         super.onActivityCreated(savedInstanceState)
         initializePresenter()
         initializeViews()
-        //startReading()
+        startReading()
     }
 
     private fun unpackExtras() {
@@ -122,9 +111,9 @@ class ReadingView : BaseMVPFragment<ReadingContract.Presenter>(), ReadingContrac
         data.add(byte.toChar())
         readingData.add(byte.toChar())
         tv_data_read.append(byte.toChar().toString())
-        println("MESSAGES Read ${byte.toChar()} -> $byte ")
-        //handleTimeReading()
-        initTimeWrite()
+        logI("${byte.toChar()} -> $byte ", customTag = Const.Logging.RECEIVED)
+        handleTimeReading()
+        //initTimeWrite()
         //handleParameterReading()
     }
 
@@ -190,12 +179,21 @@ class ReadingView : BaseMVPFragment<ReadingContract.Presenter>(), ReadingContrac
         if ((data.contains(SECOND_LINE_TOKEN) || data.contains(SECOND_LINE_TOKEN_OTHER)) && !isReadingData) {
             data.clear()
             isReadingData = true
-            thread {
-                setDeviceTime(time, deviceDate)
-            }
+            presenter.setTimeDate(time, deviceDate)
+        }
+        if (isReadingData && data.isNotEmpty()) {
+            presenter.setData(data)
+            data.clear()
         }
 
     }
+
+    override fun onTimeWriteResult(isSuccessful: Boolean) {
+        if (isSuccessful) toast(getString(R.string.setting_time_is_impossible))
+        else toast(getString(R.string.time_changed))
+
+    }
+
 
     override fun displayTimeData(time: String) {
         tv_current_time.text = String.format(getString(R.string.device_time_s), time)
@@ -205,189 +203,6 @@ class ReadingView : BaseMVPFragment<ReadingContract.Presenter>(), ReadingContrac
         time = DeviceTime(hourOfDay, minute, seconds)
         deviceDate = DeviceDate(date_picker.year, date_picker.month, date_picker.dayOfMonth)
         startReading()
-    }
-
-    private fun setDeviceTime(time: DeviceTime, deviceDate: DeviceDate) {
-        val timeDate = TimeDate()
-        val year = (deviceDate.year % 100).toByte()
-        with(timeDate) {
-            god = ((year / 10) shl 4).toByte()
-            god = god or ((year % 10).toByte())
-
-            dat = ((deviceDate.day / 10) shl 4).toByte()
-            dat = dat or ((deviceDate.day % 10).toByte())
-
-            mje = (((deviceDate.month + 1) / 10) shl 4).toByte()
-            mje = mje or (((deviceDate.month + 1) % 10).toByte())
-
-            sat = ((time.hours / 10) shl 4).toByte()
-            sat = sat or ((time.hours % 10).toByte())
-
-            min = ((time.minutes / 10) shl 4).toByte()
-            min = min or ((time.minutes % 10).toByte())
-
-            sek = ((time.seconds / 10) shl 4).toByte()
-            sek = sek or (time.seconds % 10).toByte()
-            val cal = GregorianCalendar(deviceDate.year, deviceDate.month, deviceDate.day - 1)
-
-            dan = cal.get(GregorianCalendar.DAY_OF_WEEK).toByte()
-            writeTime(timeDate)
-        }
-    }
-
-    private fun writeTime(timeDate: TimeDate) {
-        val timeString = String.format(
-            Const.Data.TIME_FORMAT,
-            timeDate.sek,
-            timeDate.min,
-            timeDate.sat,
-            timeDate.dan
-        )
-        val timeDateString = String.format(
-            Const.Data.TIME_DATE_FORMAT,
-            timeDate.sek,
-            timeDate.min,
-            timeDate.sat,
-            timeDate.dan,
-            timeDate.dat,
-            timeDate.mje,
-            timeDate.god
-        )
-
-        if (!waitAnswer(timeString)) println("Set in receiver mode!")
-        else println("Time set!")
-    }
-
-    private fun waitAnswer(time: String): Boolean {
-        var isSuccessful = false
-        loop@ for (i in 1..3) {
-            sendStringToDevice(time)
-            readMessageData = DataRXMessage()
-            if (waitMessage()) {
-                when (readMessageData.status) {
-                    ACK -> {
-                        isSuccessful = true
-                        break@loop
-                    }
-                    NAK -> continue@loop
-                    COMPLETE -> break@loop
-
-                }
-            }
-
-        }
-        readMessageData = DataRXMessage()
-        return isSuccessful
-    }
-
-    private fun waitMessage(): Boolean {
-        val timeOut = System.currentTimeMillis() + 1500
-        do {
-            if (System.currentTimeMillis() > timeOut) {
-                println("Timed out!")
-                return false
-            }
-
-        } while (!endOfMessage())
-        return true
-    }
-
-    private fun endOfMessage(): Boolean {
-        while (true) {
-            if (data.isNotEmpty()) {
-                for (dataByte in data) {
-                    if (readMessageData.status == ETX || readMessageData.status == EOT) {
-                        readMessageData.bcc = readMessageData.bcc xor dataByte.toByte()
-                        if (readMessageData.bcc == 0.toByte()) {
-                            readMessageData.status = COMPLETE
-                        } else {
-                            readMessageData.proterr = 0xCC.toByte()
-                        }
-                        return true
-                    } else {
-                        if (readMessageData.status == SOH || readMessageData.status == STX)
-                            readMessageData.bcc = readMessageData.bcc xor dataByte.toByte()
-
-                        when (dataByte.toByte()) {
-                            SOH -> {
-                                if (readMessageData.count == 0)
-                                    readMessageData.status = SOH
-                                else
-                                    readMessageData.proterr = SOH
-                                readMessageData.buffer[readMessageData.count++] = dataByte.toByte()
-                            }
-                            STX -> {
-                                if (readMessageData.status == SOH || readMessageData.count == 0)
-                                    readMessageData.status = STX
-                                else
-                                    readMessageData.proterr = STX
-                                readMessageData.buffer[readMessageData.count++] = dataByte.toByte()
-                            }
-                            0x0D.toByte() -> {
-                                readMessageData.buffer[readMessageData.count++] = dataByte.toByte()
-                                if (readMessageData.type == 0.toByte()) readMessageData.crlf = 0x0D
-                            }
-                            0x0A.toByte() -> {
-                                readMessageData.buffer[readMessageData.count++] = dataByte.toByte()
-                                if (readMessageData.type == 0.toByte()) {
-                                    if (readMessageData.crlf == 0x0D.toByte())
-                                        readMessageData.crlf = 0x0A
-                                    readMessageData.status = COMPLETE
-                                    return true
-                                }
-                            }
-                            ETX -> {
-                                readMessageData.status = ETX
-                            }
-                            EOT -> {
-                                readMessageData.status = EOT
-                            }
-                            ACK -> {
-                                readMessageData.status = ACK
-                                return true
-                            }
-                            NAK -> {
-                                readMessageData.status = NAK
-                                return true
-                            }
-                            else -> {
-                                readMessageData.buffer[readMessageData.count++] = dataByte.toByte()
-                                if (readMessageData.count > 2048 * 4) {
-                                    readMessageData.proterr = 0x55
-                                    return false
-                                }
-                            }
-
-                        }
-                    }
-                }
-            }
-
-
-            Thread.sleep(700)
-        }
-    }
-
-
-    private fun sendStringToDevice(time: String) {
-        var j = 0
-        val messageSendData = DataTXTMessage()
-        if (time.isNotEmpty()) {
-            messageSendData.buffer[j++] = SOH
-            for (char in time) {
-                messageSendData.buffer[j++] = char.toByte()
-                messageSendData.bcc = messageSendData.bcc xor char.toByte()
-            }
-            messageSendData.buffer[j++] = ETX
-            messageSendData.bcc = messageSendData.bcc xor ETX
-            messageSendData.buffer[j++] = messageSendData.bcc
-            messageSendData.count = j
-
-            CommunicationUtil.writeToSocket(
-                socket,
-                messageSendData.buffer.take(messageSendData.count).toByteArray()
-            )
-        }
     }
 
     override fun onError(throwable: Throwable) {
