+++ app/src/main/res/layout/fragment_reading.xml
@@ -15,6 +15,7 @@
             android:layout_width="match_parent"
             android:layout_height="0dp"
             android:layout_weight="3"
+            android:background="@color/light_gray"
             android:gravity="center_horizontal"
             android:orientation="vertical">
