+++ app/src/main/res/values/colors.xml
@@ -9,4 +9,5 @@
     <color name="white">#FFFFFF</color>
 
     <color name="dark_gray">#3e3e3e</color>
+    <color name="light_gray">#9e9e9e</color>
 </resources>
